sample-plugin.go: Using the instanceManager
This patch shows how to use the instanceManager.
It uses the CQL go driver instead of the http one.

The datasource would have an open connection that it will reuse. The
connection that will be stored in the instance setting.

The instanceSetting  will be passed to the query method.

This patch also show how to get information from the datasource
configuration.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>